updated collection_processor
