+++ conf/thredds-extract-assets-2.yaml
@@ -13,6 +13,5 @@ outputs:
       host: http://127.0.0.1/stac
     collection:
       name: CMIP6
-      description: The WCRP Coupled Model Intercomparison Project, Phase 6 (CMIP6), was a global climate model intercomparison project, coordinated by PCMDI (Program For Climate Model Diagnosis and Intercomparison) on behalf of the World Climate Research Program (WCRP) and provided input for the Intergovernmental Panel on Climate Change (IPCC) 6th Assessment Report (AR6).The CMIP6 archive is managed via the Earth System Grid Federation, a globally distributed archive, with various gateways with advanced faceted search capabilities provided by a number of participating organisations. Full details are available from the PCMDI CMIP6 pages (see linked documentation on this record).
 logging:
     level: INFO
